requirements.txt
streamlit
pandas
openpyxl
st-gsheets-connection
fpdf2>=2.7.8
requests
